move dev/vagrant-config/laravel/.env to .env.example add FEATURE vars

+++ dev/vagrant-config/scripts/echo.sh
@@ -4,7 +4,9 @@ echo " "
 echo "ECHO"
 echo " "
 
-cp /vagrant/dev/vagrant-config/laravel/laravel-echo-server.json /vagrant/laravel-echo-server.json
+if [ ! -e /vagrant/laravel-echo-server.json ]; then
+    cp /vagrant/dev/vagrant-config/laravel/laravel-echo-server.json /vagrant/laravel-echo-server.json
+fi
 
 systemctl enable laravel-echo-server
 systemctl start laravel-echo-server
